+++ src/python/gennylib/genny_runner.py
@@ -1,0 +1,5 @@
+def main_genny_runner():
+    print("hello world")
+
+if __name__ == '__main__':
+    genny_main()
